% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vert_norm.R
\name{add_vert_norm}
\alias{add_vert_norm}
\title{Add violin style elements to a \code{plotly} object}
\usage{
add_vert_norm(
  p,
  x = NULL,
  y = NULL,
  ...,
  sd,
  link = c("identity", "log", "logit"),
  delta,
  data = NULL,
  inherit = TRUE,
  name,
  prob = 0.95,
  step = 0.01
)
}
\arguments{
\item{p}{a plotly object}

\item{x}{the x variable.}

\item{y}{the variable median on the natural scale.}

\item{...}{Arguments (i.e., attributes) passed along to the trace \code{type}.
See \code{\link[plotly:schema]{schema()}} for a list of acceptable attributes for a given trace \code{type}
(by going to \code{traces} -> \code{type} -> \code{attributes}). Note that attributes
provided at this level may override other arguments
(e.g. \code{plot_ly(x = 1:10, y = 1:10, color = I("red"), marker = list(color = "blue"))}).}

\item{sd}{the variable of the standard error on the link scale.}

\item{link}{the link between the natural scale and the link scale.
Defaults to \code{"identity"}.}

\item{data}{A data frame (optional) or \link[crosstalk:SharedData]{crosstalk::SharedData} object.}

\item{inherit}{inherit attributes from \code{\link[plotly:plot_ly]{plot_ly()}}?}

\item{name}{Optional name of the trace for the legend.}

\item{step}{the step size between consecutive bands.
The function adds all bands with coverage \code{max_prob - i * step} for all
positive integer values \code{i} resulting in a positive coverage.
Defaults to \code{0.05}.}
}
\description{
Add violin style elements to a \code{plotly} object
}
\seealso{
Other plotly add-ons: 
\code{\link{add_classification}()},
\code{\link{add_fan}()},
\code{\link{reference_shape}()},
\code{\link{reference_text}()}
}
\concept{plotly add-ons}
