% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fan.R
\name{calculate_fan}
\alias{calculate_fan}
\title{calculate confidence intervals for different coverages}
\usage{
calculate_fan(x, fit, fit_se, coverage = c(0.3, 0.6, 0.9))
}
\arguments{
\item{x}{a \code{data.frame}}

\item{fit}{the variable from \code{x} that contains the fitted value}

\item{fit_se}{the variable from \code{x} that contains the standard error of the
fitted value}

\item{coverage}{a vector with coverage values}
}
\value{
the original data.frame with extra variables \code{lcl}, \code{ucl} and
\code{coverage}.
}
\description{
calculate confidence intervals for different coverages
}
