% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_breaks.R
\name{change_breaks}
\alias{change_breaks}
\title{Logarithmic breaks for changes}
\usage{
change_breaks(n = 2, extra = NULL)
}
\arguments{
\item{n}{the number of breaks on either side of the reference}

\item{extra}{An optional vector of additional breaks.
The function always appends these breaks.
Use this option when you want to force this values to be a part of the
breaks.}
}
\description{
Breaks a set of pretty breaks for changes.
}
\seealso{
Other utils: 
\code{\link{change_labels}()},
\code{\link{is_effectclass}()},
\code{\link{unlist}()}
}
\concept{utils}
