% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_breaks.R
\name{change_labels}
\alias{change_labels}
\title{Display logarithmic changes as percentage}
\usage{
change_labels(x)
}
\arguments{
\item{x}{the logarithmic changes}
}
\description{
Display logarithmic changes as percentage
}
