% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_effect.R
\name{class_labels}
\alias{class_labels}
\title{Return a standardised set of labels for the classification}
\usage{
class_labels(lang = c("en", "nl"), detailed = TRUE, signed = TRUE)
}
\arguments{
\item{lang}{The language.
Currently available are \code{"en"} (English) and \code{"nl"} (Dutch).
Defaults to \code{"en"}.
Please create an issue at https://github.com/inbo/effectclass/issues if you
have suggestions for more languages.}

\item{detailed}{\code{TRUE} indicates a detailed \code{\link[=classification]{classification()}};
\code{FALSE} a \code{\link[=coarse_classification]{coarse_classification()}}.
Defaults to \code{TRUE}.}

\item{signed}{\code{TRUE} indicates a signed classification;
\code{FALSE} a classification with \code{\link[=remove_sign]{remove_sign()}}.
Defaults to \code{TRUE}.}
}
\description{
Return a standardised set of labels for the classification
}
\seealso{
Other display: 
\code{\link{format_ci}()}
}
\concept{display}
