% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Classify effects by comparing the confidence intervals with a reference and thresholds}
\usage{
classification(lcl, ucl, threshold, reference = 0)
}
\arguments{
\item{lcl}{a vector of lower confidence limits}

\item{ucl}{a vector of upper confidence limits}

\item{threshold}{a vector of either 1 or 2 thresholds. A single threshold will be transformed into \code{c(-abs(threshold), abs(threshold))}.}

\item{reference}{the null hypothesis}
}
\description{
\itemize{
\item \code{++} \strong{strong positive effect}: max(threshold) < lcl
\item \code{+} \strong{positive effect}: reference < lcl < max(threshold) and max(threshold) < ucl
\item \code{+~} \strong{moderate positive effect}: reference < lcl and ucl < max(threshold)
\item \code{~} \strong{no effect}: min(threshold) < lcl < reference and reference < ucl < max(threshold)
\item \code{-~} \strong{moderate negative effect}: min(threshold) < lcl and ucl < reference
\item \code{-} \strong{negative effect}: lcl < min(threshold) and min(threshold) < ucl < reference
\item \code{--} \strong{strong negative effect}: ucl < min(threshold)
\item \code{?+} \strong{potential positive effect}: min(threshold) < lcl < reference and max(threshold) < ucl
\item \code{?-} \strong{potential negative effect}: lcl < min(threshold) and reference < ucl < max(threshold)
\item \code{?} \strong{unknown effect}: lcl < min(threshold) and max(threshold) < ucl
}
}
