% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_effectclass.R
\name{is_effectclass}
\alias{is_effectclass}
\title{Check if an object is a valid effectclass object}
\usage{
is_effectclass(x, message = c("none", "warning", "error"))
}
\arguments{
\item{x}{the object to test}

\item{message}{what to do when the object is not a valid effectclass object. "none": return FALSE with a message. "warning": return FALSE with a \code{warning()}. "error": return an error.}
}
\value{
a single TRUE or FALSE value
}
\description{
Check if an object is a valid effectclass object
}
