% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_sign.R
\name{remove_sign}
\alias{remove_sign}
\title{Remove the sign of a classifaction}
\usage{
remove_sign(classification)
}
\arguments{
\item{classification}{The classification}
}
\description{
\itemize{
\item \code{**} \strong{strong effect}: \code{++} or \code{--}
\item \code{*} \strong{effect}: \code{+} or \code{-}
\item \code{*~} \strong{moderate effect}: \code{+~} or \code{-~}
\item \code{~} \strong{no effect}: \code{~}
\item \code{?+} \strong{potential effect}: \code{?+} or \code{?-}
\item \code{?} \strong{unknown effect}: \code{?}
}
}
\seealso{
Other classification: \code{\link{classification}},
  \code{\link{coarse_classification}}
}
\concept{classification}
