% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_classification.R
\name{simple_classification}
\alias{simple_classification}
\title{Use a lower scale classification}
\usage{
simple_classification(classification)
}
\arguments{
\item{classification}{The classification}
}
\description{
\code{simplify_classification()} reduces the 10 scales from \code{classification()}
to the 4 scales below.
\itemize{
\item \code{+} \strong{positive effect}: reference < lcl
\item \code{~} \strong{no effect}: min(threshold) < lcl < reference and reference < ucl < max(threshold)
\item \code{-} \strong{negative effect}: ucl < reference
\item \code{?} \strong{unknown effect}: lcl < min(threshold) or max(threshold) < ucl
}

\code{simplify_classification()} reduces the 6 scales from
\code{remove_sign(classification())} into 3 scales.
}
