% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{unlist}
\alias{unlist}
\title{Flatten Lists}
\usage{
unlist(x, recursive = TRUE, use.names = TRUE)
}
\arguments{
\item{x}{an \R object, typically a list or vector.}

\item{recursive}{logical.  Should unlisting be applied to list
    components of \code{x}?}

\item{use.names}{logical.  Should names be preserved?}
}
\description{
Flatten Lists
}
\seealso{
base::unlist

Other utils: 
\code{\link{change_breaks}()},
\code{\link{change_labels}()},
\code{\link{is_effectclass}()}
}
\concept{utils}
