% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_classification.R
\name{add_classification}
\alias{add_classification}
\title{Add a point symbol with classification to a \code{plotly} object}
\usage{
add_classification(
  p,
  x = NULL,
  y = NULL,
  ...,
  data = NULL,
  inherit = TRUE,
  sd,
  lcl,
  ucl,
  threshold,
  reference = 0,
  prob = 0.95,
  link = c("identity", "log", "logit"),
  size = 20
)
}
\arguments{
\item{p}{a plotly object}

\item{x}{the x variable.}

\item{y}{the y variable.}

\item{...}{Arguments (i.e., attributes) passed along to the trace \code{type}.
See \code{\link[plotly:schema]{schema()}} for a list of acceptable attributes for a given trace \code{type}
(by going to \code{traces} -> \code{type} -> \code{attributes}). Note that attributes
provided at this level may override other arguments
(e.g. \code{plot_ly(x = 1:10, y = 1:10, color = I("red"), marker = list(color = "blue"))}).}

\item{data}{A data frame (optional) or \link[crosstalk:SharedData]{crosstalk::SharedData} object.}

\item{inherit}{inherit attributes from \code{\link[plotly:plot_ly]{plot_ly()}}?}

\item{sd}{the variable of the standard error on the link scale.}

\item{lcl}{A vector of lower confidence limits.}

\item{ucl}{A vector of upper confidence limits.}

\item{threshold}{A vector of either 1 or 2 thresholds.
A single threshold will be transformed into
\code{reference + c(-abs(threshold), abs(threshold))}.}

\item{reference}{The null hypothesis.
Defaults to 0.}

\item{prob}{The coverage of the confidence interval when calculated from the
mean \code{y} and standard error \code{sd}.
Note that the function assumes a normal distribution at the \code{link} scale.}

\item{link}{the link between the natural scale and the link scale.
Defaults to \code{"identity"}.}

\item{size}{Size of the point symbol.}
}
\description{
See \code{classification()} for an explication on how the classification is done.
}
\seealso{
Other plotly add-ons: 
\code{\link{add_fan}()}
}
\concept{plotly add-ons}
