% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fan.R
\name{add_fan}
\alias{add_fan}
\title{Add a fan plot to a \code{plotly} object}
\usage{
add_fan(
  p,
  x = NULL,
  y = NULL,
  ...,
  sd,
  link = c("identity", "log", "logit"),
  max_prob = 0.9,
  step = 0.05,
  fillcolor = "blue",
  data = NULL,
  inherit = TRUE
)
}
\arguments{
\item{p}{a plotly object}

\item{x}{the x variable.}

\item{y}{the variable median on the natural scale.}

\item{...}{Arguments (i.e., attributes) passed along to the trace \code{type}.
See \code{\link[plotly:schema]{schema()}} for a list of acceptable attributes for a given trace \code{type}
(by going to \code{traces} -> \code{type} -> \code{attributes}). Note that attributes
provided at this level may override other arguments
(e.g. \code{plot_ly(x = 1:10, y = 1:10, color = I("red"), marker = list(color = "blue"))}).}

\item{sd}{the variable of the standard error on the link scale.}

\item{link}{the link between the natural scale and the link scale.
Defaults to \code{"identity"}.}

\item{max_prob}{The coverage of the widest band.
Defaults to \code{0.9}.}

\item{step}{the step size between consecutive bands.
The function adds all bands with coverage \code{max_prob - i * step} for all
positive integer values \code{i} resulting in a positive coverage.
Defaults to \code{0.05}.}

\item{fillcolor}{The fill colour of the fan.
Defaults to \code{"blue"}.}

\item{data}{A data frame (optional) or \link[crosstalk:SharedData]{crosstalk::SharedData} object.}

\item{inherit}{inherit attributes from \code{\link[plotly:plot_ly]{plot_ly()}}?}
}
\description{
A fan plot consist of a set of transparent ribbons each representing a
different coverage of the uncertainty around an estimate.
The coverages are based on the assumption of a normal distribution with mean
\code{link(y)} and standard error \code{sd}.
}
\examples{
set.seed(20191218)
z <- data.frame(
  year = 1990:2019, dx = rnorm(30, sd = 0.2), s = rnorm(30, 0.5, 0.01)
)
z$index <- 3 + cumsum(z$dx)
library(plotly)
p <- plot_ly(z, x = ~year, y = ~index)
p |>
  add_fan(sd = ~s)
p |>
  add_fan(sd = ~s, step = 0.3, fillcolor = "green")
p |>
  add_fan(sd = ~s, max_prob = 0.95, link = "log")
}
\seealso{
Other plotly add-ons: 
\code{\link{add_classification}()},
\code{\link{reference_shape}()}
}
\concept{plotly add-ons}
