% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_breaks.R
\name{change_breaks}
\alias{change_breaks}
\title{Logarithmic breaks for changes}
\usage{
change_breaks(n = 2)
}
\arguments{
\item{n}{the number of breaks on either side of the reference}
}
\description{
Breaks a set of pretty breaks for changes.
}
