% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Classify Effects by Comparing the Confidence Intervals with a Reference and
Thresholds}
\usage{
classification(lcl, ucl, threshold, reference = 0)
}
\arguments{
\item{lcl}{A vector of lower confidence limits.}

\item{ucl}{A vector of upper confidence limits.}

\item{threshold}{A vector of either 1 or 2 thresholds.
A single threshold will be transformed into
\code{reference + c(-abs(threshold), abs(threshold))}.}

\item{reference}{The null hypothesis.
Defaults to 0.}
}
\description{
\itemize{
\item \verb{++} \strong{strong positive effect}: \code{max(threshold) < lcl}
\item \code{+} \strong{positive effect}: \verb{reference < lcl < max(threshold)} and
\code{max(threshold) < ucl}
\item \verb{+~} \strong{moderate positive effect}: \code{reference < lcl} and
\code{ucl < max(threshold)}
\item \code{~} \strong{no effect}: \verb{min(threshold) < lcl < reference} and
\verb{reference < ucl < max(threshold)}
\item \verb{-~} \strong{moderate negative effect}: \code{min(threshold) < lcl} and
\code{ucl < reference}
\item \code{-} \strong{negative effect}: \code{lcl < min(threshold)} and
\verb{min(threshold) < ucl < reference}
\item \verb{--} \strong{strong negative effect}: \code{ucl < min(threshold)}
\item \verb{?+} \strong{potential positive effect}: \verb{min(threshold) < lcl < reference} and
\code{max(threshold) < ucl}
\item \verb{?-} \strong{potential negative effect}: \code{lcl < min(threshold)} and
\verb{reference < ucl < max(threshold)}
\item \verb{?} \strong{unknown effect}: \code{lcl < min(threshold)} and \code{max(threshold) < ucl}
}
}
\seealso{
Other classification functions: 
\code{\link{coarse_classification}()},
\code{\link{remove_sign}()}
}
\concept{classification functions}
