% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_shape.R
\name{reference_shape}
\alias{reference_shape}
\title{Create \code{plotly}references
Returns a list shapes you can pass to the \code{shapes} argument of
\code{plotly::layout()}}
\usage{
reference_shape(threshold, reference = 0, colour = "black", horizontal = TRUE)
}
\arguments{
\item{threshold}{A vector of either 1 or 2 thresholds.
A single threshold will be transformed into
\code{reference + c(-abs(threshold), abs(threshold))}.}

\item{reference}{The null hypothesis.
Defaults to 0.}

\item{colour}{The colour for the references.
Defaults to \code{"black"}.}

\item{horizontal}{Display horizontal reference when \code{TRUE} (default).
Display vertical reference when \code{FALSE}.}
}
\description{
Create \code{plotly}references
Returns a list shapes you can pass to the \code{shapes} argument of
\code{plotly::layout()}
}
\examples{
z <- data.frame(
  effect = factor(
    1:10,
    labels = c("unknown\neffect", "potential\npositive\neffect",
             "potential\nnegative\neffect", "no effect", "positive\neffect",
             "negative\neffect", "moderate\npositive\neffect",
             "moderate\nnegative\neffect", "strong\npositive\neffect",
             "strong\nnegative\neffect")
  ),
  estimate = c( 0,  0,    0,   0,   1,   -1,   0.5, -0.5, 1.5, -1.5),
  lcl =      c(-2, -0.9, -2,  -0.9, 0.1, -2,   0.1, -0.9, 1.1, -2),
  ucl =      c( 2,  2,    0.9, 0.9, 2,   -0.1, 0.9, -0.1, 2,   -1.1)
)
library(plotly)
z$error_up <- z$ucl - z$estimate
z$error_down <- z$estimate - z$lcl
plot_ly(
  z, y = ~effect, x = ~estimate, type = "scatter", mode = "none",
  error_x = list(type = "data", array = ~error_up, arrayminus = ~error_down)
) |>
  add_classification(lcl = ~lcl, ucl = ~ucl, threshold = 1) |>
  layout(
    shapes = reference_shape(threshold = 1, horizontal = FALSE)
  )
plot_ly(
  z, y = ~effect, x = ~estimate, type = "scatter", mode = "none",
  error_x = list(type = "data", array = ~error_up, arrayminus = ~error_down)
) |>
  add_classification(
  lcl = ~lcl, ucl = ~ucl, threshold = 1, detailed = FALSE
 ) |>
  layout(
    shapes = reference_shape(threshold = 1, horizontal = FALSE)
  )
plot_ly(
  z, y = ~effect, x = ~estimate, type = "scatter", mode = "none",
  error_x = list(type = "data", array = ~error_up, arrayminus = ~error_down)
) |>
  add_classification(
    lcl = ~lcl, ucl = ~ucl, threshold = 1, signed = FALSE
   ) |>
  layout(
    shapes = reference_shape(threshold = 1, horizontal = FALSE)
  )
plot_ly(
  z, y = ~effect, x = ~estimate, type = "scatter", mode = "none",
  error_x = list(type = "data", array = ~error_up, arrayminus = ~error_down)
) |>
  add_classification(
    lcl = ~lcl, ucl = ~ucl, threshold = 1, detailed = FALSE, signed = FALSE
  ) |>
  layout(
    shapes = reference_shape(threshold = 1, horizontal = FALSE)
  )
}
\seealso{
Other plotly add-ons: 
\code{\link{add_classification}()},
\code{\link{add_fan}()}
}
\concept{plotly add-ons}
