% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_sign.R
\name{remove_sign}
\alias{remove_sign}
\title{Remove the sign of a classifaction}
\usage{
remove_sign(classification)
}
\arguments{
\item{classification}{The classification}
}
\description{
\itemize{
\item \verb{**} \strong{strong effect}: \verb{++} or \verb{--}
\item \code{*} \strong{effect}: \code{+} or \code{-}
\item \verb{*~} \strong{moderate effect}: \verb{+~} or \verb{-~}
\item \code{~} \strong{no effect}: \code{~}
\item \verb{?+} \strong{potential effect}: \verb{?+} or \verb{?-}
\item \verb{?} \strong{unknown effect}: \verb{?}
}
}
\seealso{
Other classification: 
\code{\link{classification}()},
\code{\link{coarse_classification}()}
}
\concept{classification}
