% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_effect.R
\name{stat_effect}
\alias{stat_effect}
\title{Display points with classified effect}
\usage{
stat_effect(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...,
  threshold, reference = 0, detailed = TRUE, signed = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{threshold}{A vector of either 1 or 2 thresholds.
A single threshold will be transformed into
\code{reference + c(-abs(threshold), abs(threshold))}.}

\item{reference}{The null hypothesis.
Defaults to 0.}

\item{detailed}{\code{TRUE} indicates a detailed \code{\link{classification}};
\code{FALSE} a \code{\link{coarse_classification}}.
Defaults to \code{TRUE}.}

\item{signed}{\code{TRUE} indicates a signed classification;
\code{FALSE} a classification with \code{\link{remove_sign}}.
Defaults to \code{TRUE}.}
}
\description{
Display points with classified effect
}
\examples{
z <- data.frame(
  effect = factor(
    1:10,
    labels = c("unknown\\neffect", "potential\\npositive\\neffect",
             "potential\\nnegative\\neffect", "no effect", "positive\\neffect",
             "negative\\neffect", "moderate\\npositive\\neffect",
             "moderate\\nnegative\\neffect", "strong\\npositive\\neffect",
             "strong\\nnegative\\neffect")
  ),
  estimate = c( 0,  0,    0,   0,   1,   -1,   0.5, -0.5, 1.5, -1.5),
  lcl =      c(-2, -0.9, -2,  -0.9, 0.1, -2,   0.1, -0.9, 1.1, -2),
  ucl =      c( 2,  2,    0.9, 0.9, 2,   -0.1, 0.9, -0.1, 2,   -1.1)
)
library(ggplot2)
ggplot(z, aes(x = effect, y = estimate, ymin = lcl, ymax = ucl)) +
  geom_hline(yintercept = c(-1, 1, 0), linetype = c(3, 3, 2)) +
  geom_errorbar() +
  stat_effect(threshold = 1, size = 3) +
  scale_effect() +
  coord_flip()
ggplot(z[3:5, ], aes(x = effect, y = estimate, ymin = lcl, ymax = ucl)) +
  geom_hline(yintercept = c(-1, 1, 0), linetype = c(3, 3, 2)) +
  geom_errorbar() +
  stat_effect(threshold = 1, size = 3) +
  scale_effect() +
  coord_flip()
ggplot(z[3:5, ], aes(x = effect, y = estimate, ymin = lcl, ymax = ucl)) +
  geom_hline(yintercept = c(-1, 1, 0), linetype = c(3, 3, 2)) +
  geom_errorbar() +
  stat_effect(threshold = 1, size = 3) +
  scale_effect(drop = TRUE) +
  coord_flip()

# plot indices
set.seed(20190521)
base_year <- 2000
n_year <- 10
z <- data.frame(
  dt = seq_len(n_year),
  change = rnorm(n_year, sd = 0.2),
  sd = rnorm(n_year, mean = 0.1, sd = 0.01)
)
z$index <- cumsum(z$change)
z$lcl <- qnorm(0.025, z$index, z$sd)
z$ucl <- qnorm(0.975, z$index, z$sd)
z$year <- base_year + z$dt
th <- 0.25
ref <- 0
ggplot(z, aes(x = year, y = index, ymin = lcl, ymax = ucl, sd = sd)) +
  geom_hline(yintercept = c(ref, -th, th), linetype = c(2, 3, 3)) +
  geom_errorbar() +
  geom_line() +
  stat_effect(threshold = th, reference = ref, size = 3) +
  scale_effect()

# plot pairwise differences
change_set <- function(z, base_year) {
  n_year <- max(z$dt)
  total_change <- lapply(
    seq_len(n_year) - 1,
    function(i) {
      if (i > 0) {
        y <- tail(z, -i)
      } else {
        y <- z
      }
      data.frame(
        from = base_year + i,
        to = base_year + y$dt,
        total = cumsum(y$change),
        sd = sqrt(cumsum(y$sd ^ 2))
      )
    }
  )
  total_change <- do.call(rbind, total_change)
  total_change <- rbind(
    total_change,
    data.frame(
      from = total_change$to,
      to = total_change$from,
      total = -total_change$total,
      sd = total_change$sd
    )
  )
  total_change$lcl <- qnorm(0.025, total_change$total, total_change$sd)
  total_change$ucl <- qnorm(0.975, total_change$total, total_change$sd)
  return(total_change)
}
ggplot(change_set(z, base_year),
       aes(x = from, y = to, ymin = lcl, ymax = ucl)) +
  stat_effect(threshold = th, reference = ref, aes(colour = total),
              size = 3) +
  scale_colour_gradient2() +
  scale_effect()
}
\seealso{
\code{\link{classification}}

Other ggplot2: \code{\link{scale_effect}},
  \code{\link{stat_fan}}
}
\concept{ggplot2}
