% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fan.R
\docType{data}
\name{stat_fan}
\alias{stat_fan}
\alias{StatFan}
\title{Display a fan plot}
\usage{
stat_fan(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  geom = "ribbon",
  ...,
  fine = FALSE,
  link = c("identity", "log", "logit")
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{geom}{Use a different \code{geom} than the default \code{"ribbon"}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{fine}{a logical value.
\code{TRUE} displays coverages from 10\\% to 90\\% in steps of 10\\%.
\code{FALSE} displays coverages from 30\\% to 90\\% in steps of 30\\%.
Defaults to \code{FALSE}}

\item{link}{the link function to apply on the \code{y} before calculating the
coverage intervals.
Note that \code{link_sd} is the standard error on the link scale,
while \code{y} is on the natural scale.
Defaults to \code{'identify'} which implies no transformation (\code{link(y) == y}).
Other options are \code{'log'} and \code{'logit'}.}
}
\description{
A fan plot consist of a set of transparent ribbons each representing a
different coverage of the uncertainty around an estimate.
The coverages are based on the assumption of a normal distribution with mean
\code{link(y)} and standard error \code{link_sd}.
}
\examples{
set.seed(20191218)
z <- data.frame(
  year = 1990:2019,
  dx = rnorm(30, sd = 0.2),
  s = rnorm(30, 0.5, 0.01)
 )
z$index <- 3 + cumsum(z$dx)
library(ggplot2)
ggplot(z, aes(x = year, y = index, link_sd = s)) + stat_fan()
ggplot(z, aes(x = year, y = index, link_sd = s)) + stat_fan() + geom_line()
ggplot(z, aes(x = year, y = index, link_sd = s)) + stat_fan(fine = TRUE)
ggplot(z, aes(x = year, y = exp(index), link_sd = s)) +
  stat_fan(link = "log") + geom_line()
ggplot(z, aes(x = year, y = plogis(index), link_sd = s)) +
  stat_fan(link = "logit") + geom_line()
ggplot(z, aes(x = year, y = index, link_sd = s)) + stat_fan(geom = "rect")
ggplot(z, aes(x = year, y = index, link_sd = s)) + stat_fan(geom = "bar")
ggplot(z, aes(x = year, y = index, link_sd = s)) +
  stat_fan(geom = "errorbar")
ggplot(z, aes(x = year, y = index, link_sd = s)) +
  stat_fan(geom = "linerange")
ggplot(z, aes(x = year, y = index, link_sd = s)) +
  stat_fan(geom = "pointrange")

z <- expand.grid(year = 1990:2019, category = c("A", "B"))
z$dx <- rnorm(60, sd = 0.1)
z$index <- rep(c(0, 2), each = 30) + cumsum(z$dx)
z$s <- rnorm(60, rep(c(0.5, 1), each = 30), 0.05)
ggplot(z, aes(x = year, y = index, link_sd = s)) + stat_fan() + geom_line() +
  facet_wrap(~category)
ggplot(z, aes(x = year, y = index, link_sd = s)) +
  stat_fan(aes(fill = category)) + geom_line(aes(colour = category))
}
\seealso{
Other ggplot2 add-ons: 
\code{\link{scale_effect}()},
\code{\link{stat_effect}()}
}
\concept{ggplot2 add-ons}
\keyword{datasets}
